moved function to util.jsx file
